+++ CHANGELOG.md
@@ -6,8 +6,7 @@ All notable changes for each version of this project will be documented in this 
     - **Breaking Change** `combo.value` is now only a getter.
 - `igxDropDown`
     - `IgxDropDownItemBase` and it's descendants (of which `IgxDropDownItem`) have had their `isSelected` and `isFocused` properties **deprecated**. Instead, use `selected` and `focused` properties.
-    - **Breaking Change** the `index` property representing the visible index of the rendered item in the virtualized dropdown container is now called `itemIndex`.
-    - Added an input for the data index (such as the one coming from ngFor) of the `IgxDropDownItem` component.
+    - Added an `@Input` for the `index` property (such as the one coming from ngFor) of the `IgxDropDownItem` component. This **deprecates** the automatic index calculation.
     ```html
         <igx-drop-down>
             <igx-drop-down-item *ngFor="let item of items; let i = index" [index]="i">